install.sh: Improve default $PREFIX behaviour
As before we default to /usr/local but now we switch to ~/.local if
/usr/local if not accessible instead of prompting for sudo. Instead
of having an exception for macOS/arm64.

Of course if a user wants to install in /usr/local anyway, they can
still pass --prefix /usr/local to force installation into /usr/local
and prompt for sudo.